feat(media): add Cloudinary upload module

+++ backend/internal/shared/cloudinary/client.go
@@ -21,7 +21,7 @@ type UploadResult struct {
 
 	ResourceType string
 
-	Bytes int64
+	Bytes int
 
 	Format string
 }
